+++ packages/ngw-kit/src/adapters/createGeoJsonAdapter.ts
@@ -62,8 +62,6 @@ export async function createGeoJsonAdapter(
       filters,
       connector,
       ...opt,
-      // strict restriction on loading data from large layers
-      limit: opt?.limit !== undefined ? opt.limit : 3000,
     });
     return await _dataPromise;
   };
@@ -286,7 +284,7 @@ export async function createGeoJsonAdapter(
         return `POLYGON((${polygon.join(', ')}))`;
       }
     }
-  };
+  }
 
   return NgwGeoJsonAdapter;
 }